# SmartPowerElectricAPI/bin/Release/net8.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "SmartPowerElectricAPI, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null",
    "version": "1.0"
  },
  "paths": {
...
    "/api/Orden/create/{idProyecto}": {
      "post": {
        "tags": [
          "Orden"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "idProyecto",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/OrdenDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Orden/{id}": {
      "delete": {
        "tags": [
          "Orden"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "put": {
        "tags": [
          "Orden"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/OrdenDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "get": {
        "tags": [
          "Orden"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Orden/asociarOrdenTrabajador": {
      "post": {
        "tags": [
          "Orden"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/OrdenTrabajadorPar"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/Proyecto/create": {
      "post": {
        "tags": [
          "Proyecto"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/ProyectoDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Proyecto/{id}": {
      "delete": {
        "tags": [
          "Proyecto"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "put": {
        "tags": [
          "Proyecto"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/ProyectoDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "get": {
        "tags": [
          "Proyecto"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "FacturaDTO": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "montoACobrar": {
          "format": "double",
          "type": "number"
        },
        "numeroFactura": {
          "type": "string"
        },
        "emailEnviado": {
          "type": "boolean"
        },
        "facturaCompletada": {
          "type": "boolean"
        },
        "idOrden": {
          "format": "int32",
          "type": "integer"
        },
        "fechaCreacion": {
          "type": "string"
        },
        "fechaEliminado": {
          "type": "string"
        },
        "eliminado": {
          "type": "boolean"
        }
      },
      "additionalProperties": false
    },
...
    "MaterialDTO": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "precio": {
          "format": "double",
          "type": "number"
        },
        "cantidad": {
          "format": "double",
          "type": "number"
        },
        "precioTotal": {
          "format": "double",
          "type": "number"
        },
        "idTipoMaterial": {
          "format": "int32",
          "type": "integer"
        },
        "tipoMaterial": {
          "type": "string"
        },
        "idUnidadMedida": {
          "format": "int32",
          "type": "integer"
        },
        "unidadMedida": {
          "type": "string"
        },
        "idOrden": {
          "format": "int32",
          "type": "integer"
        },
        "fechaCreacion": {
          "type": "string"
        },
        "fechaEliminado": {
          "type": "string"
        },
        "eliminado": {
          "type": "boolean"
        }
      },
      "additionalProperties": false
    },
    "OrdenDTO": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "numeroOrden": {
          "format": "int32",
          "type": "integer"
        },
        "ordenFinalizada": {
          "type": "boolean"
        },
        "costeManoObra": {
          "format": "double",
          "type": "number"
        },
        "cobrado": {
          "format": "double",
          "type": "number"
        },
        "horasEstimadas": {
          "format": "double",
          "type": "number"
        },
        "idProyecto": {
          "format": "int32",
          "type": "integer"
        },
        "nombreProyecto": {
          "type": "string"
        },
        "materialDTOs": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/MaterialDTO"
          }
        },
        "trabajadorDTOs": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/TrabajadorDTO"
          }
        },
        "facturaDTOs": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/FacturaDTO"
          }
        },
        "fechaCreacion": {
          "type": "string"
        },
        "fechaEliminado": {
          "type": "string"
        },
        "eliminado": {
          "type": "boolean"
        },
        "costeMateriales": {
          "format": "double",
          "type": "number"
        },
        "costeTotal": {
          "format": "double",
          "type": "number"
        },
        "faltanteCobrar": {
          "format": "double",
          "type": "number"
        }
      },
      "additionalProperties": false
    },
    "OrdenTrabajadorPar": {
      "type": "object",
      "properties": {
        "idOrden": {
          "format": "int32",
          "type": "integer"
        },
        "idTrabajador": {
          "format": "int32",
          "type": "integer"
        }
      },
      "additionalProperties": false
    },
    "ProyectoDTO": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "nombre": {
          "type": "string"
        },
        "direccion": {
          "type": "string"
        },
        "descripcion": {
          "type": "string"
        },
        "fechaInicio": {
          "type": "string"
        },
        "fechaFin": {
          "type": "string"
        },
        "idCliente": {
          "format": "int32",
          "type": "integer"
        },
        "nombreCLiente": {
          "type": "string"
        },
        "horasEstimadas": {
          "format": "double",
          "type": "number"
        },
        "fechaCreacion": {
          "type": "string"
        },
        "fechaEliminado": {
          "type": "string"
        },
        "eliminado": {
          "type": "boolean"
        },
        "finalizado": {
          "type": "boolean"
        },
        "ordenDTOs": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/OrdenDTO"
          }
        }
      },
      "additionalProperties": false
    },
...
    "TrabajadorDTO": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "nombre": {
          "type": "string"
        },
        "apellido": {
          "type": "string"
        },
        "especialidad": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "telefono": {
          "type": "string"
        },
        "direccion": {
          "type": "string"
        },
        "seguridadSocial": {
          "type": "string"
        },
        "fechaInicioContrato": {
          "type": "string"
        },
        "fechaFinContrato": {
          "type": "string"
        },
        "cobroxHora": {
          "format": "double",
          "type": "number"
        },
        "numeroCuenta": {
          "type": "string"
        },
        "enrutamiento": {
          "type": "string"
        },
        "fechaCreacion": {
          "type": "string"
        },
        "fechaEliminado": {
          "type": "string"
        },
        "eliminado": {
          "type": "boolean"
        }
      },
      "additionalProperties": false
    },
...
  }
}